requirements.txt
Telethon==1.38.1
streamlit>=1.30.0
nest-asyncio>=1.5.8
python-dotenv==1.0.1
cryptg==0.5.0.post0
